Auth Some Controllers

+++ Server/src/main/java/server/server/Controller/EventController.java
@@ -14,6 +14,7 @@ import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.validation.Errors;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -26,8 +27,12 @@ import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 import server.server.Controller.Utilities.Utility;
 import server.server.Model.Domain.Event;
+import server.server.Model.Services.IDepartmentService;
 import server.server.Model.Services.IEventService;
+import server.server.Model.Services.Impls.CustomUserDetails;
+import server.server.auth.IAuthenticationFacade;
 import server.server.utilities.Labels;
+import server.server.utilities.errors.UserErrors;
 
 /**
  *
@@ -40,18 +45,26 @@ public class EventController {
     @Autowired
     private IEventService evtService;
 
+    @Autowired
+    private IDepartmentService dptService;
+
+    @Autowired
+    private IAuthenticationFacade authenticationFacade;
+
     /**
      * Find an event by code or id
      *
      * @param params
      * @return
      */
+    @PreAuthorize("hasRole('ADMIN') or hasRole('SCHEDULEMANAGER')")
     @GetMapping(value = "/find")
     @ResponseBody
     public ResponseEntity<Event> get(@RequestParam Map<String, String> params) {
         ResponseEntity responseEntity = new ResponseEntity<>(null, null, HttpStatus.NOT_FOUND);
         Map<Labels, Object> returns = null;
         if (params.keySet().size() == 1) {
+
             if (params.containsKey("id")) {
                 Long id = Long.parseLong(params.get("id"));
                 returns = evtService.findbyId(id);
@@ -63,6 +76,17 @@ public class EventController {
             if (returns != null) {
                 Event ev = (Event) returns.get(Labels.objectReturn);
                 if (ev != null) {
+
+                    HttpHeaders headers = new HttpHeaders();
+                    ArrayList<String> errors2 = new ArrayList();
+                    CustomUserDetails user = (CustomUserDetails) authenticationFacade.getPrincipal();
+
+                    if (!dptService.validateUserDepartment(ev.getDepartment().getDepartmentId(), user)) {
+                        errors2.add(UserErrors.USR105.name());
+                        headers.add(Labels.errors.toString(), errors2.toString());
+                        return new ResponseEntity<>(null, headers, HttpStatus.UNAUTHORIZED);
+                    }
+
                     responseEntity = new ResponseEntity<>(ev, null, HttpStatus.FOUND);
                 }
             }
@@ -75,21 +99,31 @@ public class EventController {
         return responseEntity;
     }
 
+    @PreAuthorize("hasRole('ADMIN') or hasRole('SCHEDULEMANAGER')")
     @PostMapping(
             produces = MediaType.APPLICATION_JSON_VALUE,
             consumes = MediaType.APPLICATION_JSON_VALUE)
     @RequestMapping("/add")
     public ResponseEntity<Event> add(@RequestBody @Valid Event event, Errors errors) {
 
+        HttpHeaders headers = new HttpHeaders();
+        ArrayList<String> errors2 = new ArrayList();
+        CustomUserDetails user = (CustomUserDetails) authenticationFacade.getPrincipal();
+
+        if (!dptService.validateUserDepartment(event.getDepartment().getDepartmentId(), user)) {
+            errors2.add(UserErrors.USR105.name());
+            headers.add(Labels.errors.toString(), errors2.toString());
+            return new ResponseEntity<>(null, headers, HttpStatus.UNAUTHORIZED);
+        }
+
         Map<Labels, Object> returns = evtService.create(event);
-        ArrayList<String> errors2 = (ArrayList<String>) returns.get(Labels.errors);
+        errors2 = (ArrayList<String>) returns.get(Labels.errors);
         Event event2 = (Event) returns.get(Labels.objectReturn);
         ResponseEntity responseEntity;
 
         if (event2 != null) {
             responseEntity = new ResponseEntity<>(event2, null, HttpStatus.ACCEPTED);
         } else {
-            HttpHeaders headers = new HttpHeaders();
             ArrayList<String> setErrors = Utility.setErrors(errors);
             errors2.addAll(setErrors);
             headers.add(Labels.errors.name(), errors2.toString());
@@ -99,19 +133,29 @@ public class EventController {
         return responseEntity;
     }
 
+    @PreAuthorize("hasRole('ADMIN') or hasRole('SCHEDULEMANAGER')")
     @PutMapping(produces = MediaType.APPLICATION_JSON_VALUE,
             consumes = MediaType.APPLICATION_JSON_VALUE)
     @RequestMapping("/update")
     public ResponseEntity<Event> update(@RequestBody Event event, Errors errors) {
         ResponseEntity<Event> responseEntity;
+        HttpHeaders headers = new HttpHeaders();
+        ArrayList<String> errors2 = new ArrayList();
+        CustomUserDetails user = (CustomUserDetails) authenticationFacade.getPrincipal();
+
+        if (!evtService.validateUserEvent(event.getId(), user)) {
+            errors2.add(UserErrors.USR105.name());
+            headers.add(Labels.errors.toString(), errors2.toString());
+            return new ResponseEntity<>(null, headers, HttpStatus.UNAUTHORIZED);
+        }
+
         Map<Labels, Object> returns = evtService.update(event);
-        ArrayList<String> errors2 = (ArrayList<String>) returns.get(Labels.errors);
+        errors2 = (ArrayList<String>) returns.get(Labels.errors);
         Event event2 = (Event) returns.get(Labels.objectReturn);
 
         if (event2 != null) {
             responseEntity = new ResponseEntity<>(event2, null, HttpStatus.ACCEPTED);
         } else {
-            HttpHeaders headers = new HttpHeaders();
             ArrayList<String> setErrors = Utility.setErrors(errors);
             errors2.addAll(setErrors);
             headers.add(Labels.errors.name(), errors2.toString());
@@ -121,9 +165,20 @@ public class EventController {
         return responseEntity;
     }
 
+    @PreAuthorize("hasRole('ADMIN') or hasRole('SCHEDULEMANAGER')")
     @DeleteMapping
     @RequestMapping("/delete")
     public ResponseEntity<Event> delete(@RequestParam("id") long id, @RequestParam("departmentId") long departmentId) {
+
+        HttpHeaders headers = new HttpHeaders();
+        ArrayList<String> errors2 = new ArrayList();
+        CustomUserDetails user = (CustomUserDetails) authenticationFacade.getPrincipal();
+
+        if (!evtService.validateUserEvent(id, user)) {
+            errors2.add(UserErrors.USR105.name());
+            headers.add(Labels.errors.toString(), errors2.toString());
+            return new ResponseEntity<>(null, headers, HttpStatus.UNAUTHORIZED);
+        }
 
         ResponseEntity<Event> responseEntity;
         Map<Labels, Object> returns = evtService.delete(id, departmentId);
@@ -132,13 +187,13 @@ public class EventController {
         if (event != null) {
             responseEntity = new ResponseEntity<>(event, null, HttpStatus.ACCEPTED);
         } else {
-            HttpHeaders headers = new HttpHeaders();
             headers.add(Labels.errors.name(), errors.toString());
             responseEntity = (new ResponseEntity<>(event, headers, HttpStatus.NOT_MODIFIED));
         }
         return responseEntity;
     }
 
+    @PreAuthorize("hasRole('ADMIN')")
     @ResponseBody
     @GetMapping(value = "/findByDepartmentOrProgram")
     public ResponseEntity<List<Event>> findByDepartmentOrProgram(@RequestParam("departmentId") List<Long> departmentId) {
@@ -154,11 +209,20 @@ public class EventController {
         return responseEntity;
     }
 
+    @PreAuthorize("hasRole('ADMIN') or hasRole('SCHEDULEMANAGER')")
     @ResponseBody
     @GetMapping(value = "/findByDepartmentAndTypes")
     public ResponseEntity<List<Event>> findByDepartmentAndTypes(@RequestParam("departmentId") long departmentId,
             @RequestParam("type") Optional<List<String>> types) {
         ResponseEntity responseEntity = new ResponseEntity<>(null, null, HttpStatus.NOT_FOUND);
+        HttpHeaders headers = new HttpHeaders();
+        ArrayList<String> errors2 = new ArrayList();
+        CustomUserDetails user = (CustomUserDetails) authenticationFacade.getPrincipal();
+        if (!dptService.validateUserDepartment(departmentId, user)) {
+            errors2.add(UserErrors.USR105.name());
+            headers.add(Labels.errors.toString(), errors2.toString());
+            return new ResponseEntity<>(null, headers, HttpStatus.UNAUTHORIZED);
+        }
 
         List<String> types2 = null;
         if (types.isPresent() && types.isEmpty()) {
